chore: Move Select component to UI directory
