--- packages/webapp-libs/webapp-core/src/components/forms/select/index.ts
@@ -1,8 +1,0 @@
-export { Select, SelectValue, SelectGroup } from './select.component';
-export { SelectTrigger } from './selectTrigger';
-export { SelectScrollUpButton } from './selectScrollUpButton';
-export { SelectScrollDownButton } from './selectScrollDownButton';
-export { SelectContent } from './selectContent';
-export { SelectLabel } from './selectLabel';
-export { SelectItem } from './selectItem';
-export { SelectSeparator } from './selectSeparator';
