--- packages/webapp-libs/webapp-core/src/components/forms/select/select.component.tsx
@@ -1,9 +1,0 @@
-import * as SelectPrimitive from '@radix-ui/react-select';
-
-const Select = SelectPrimitive.Root;
-
-const SelectGroup = SelectPrimitive.Group;
-
-const SelectValue = SelectPrimitive.Value;
-
-export { Select, SelectGroup, SelectValue };
