--- packages/webapp-libs/webapp-core/src/components/forms/select/selectContent/index.ts
@@ -1,1 +1,0 @@
-export { SelectContent } from './selectContent.component';
