--- packages/webapp-libs/webapp-core/src/components/forms/select/selectItem/index.ts
@@ -1,1 +1,0 @@
-export { SelectItem } from './selectItem.component';
