--- packages/webapp-libs/webapp-core/src/components/forms/select/selectLabel/index.ts
@@ -1,1 +1,0 @@
-export { SelectLabel } from './selectLabel.component';
