--- packages/webapp-libs/webapp-core/src/components/forms/select/selectScrollDownButton/index.ts
@@ -1,1 +1,0 @@
-export { SelectScrollDownButton } from './selectScrollDownButton.component';
