--- packages/webapp-libs/webapp-core/src/components/forms/select/selectScrollDownButton/selectScrollDownButton.component.tsx
@@ -1,22 +1,0 @@
-import * as SelectPrimitive from '@radix-ui/react-select';
-import { ChevronDown } from 'lucide-react';
-import React from 'react';
-
-import { cn } from '../../../../lib/utils';
-
-const SelectScrollDownButton = React.forwardRef<
-  React.ElementRef<typeof SelectPrimitive.ScrollDownButton>,
-  React.ComponentPropsWithoutRef<typeof SelectPrimitive.ScrollDownButton>
->(({ className, ...props }, ref) => (
-  <SelectPrimitive.ScrollDownButton
-    ref={ref}
-    className={cn('flex cursor-default items-center justify-center py-1', className)}
-    {...props}
-  >
-    <ChevronDown className="h-4 w-4" />
-  </SelectPrimitive.ScrollDownButton>
-));
-
-SelectScrollDownButton.displayName = SelectPrimitive.ScrollDownButton.displayName;
-
-export { SelectScrollDownButton };
