--- packages/webapp-libs/webapp-core/src/components/forms/select/selectScrollUpButton/index.ts
@@ -1,1 +1,0 @@
-export { SelectScrollUpButton } from './selectScrollUpButton.component';
