--- packages/webapp-libs/webapp-core/src/components/forms/select/selectScrollUpButton/selectScrollUpButton.component.tsx
@@ -1,21 +1,0 @@
-import * as SelectPrimitive from '@radix-ui/react-select';
-import { ChevronUp } from 'lucide-react';
-import React from 'react';
-
-import { cn } from '../../../../lib/utils';
-
-const SelectScrollUpButton = React.forwardRef<
-  React.ElementRef<typeof SelectPrimitive.ScrollUpButton>,
-  React.ComponentPropsWithoutRef<typeof SelectPrimitive.ScrollUpButton>
->(({ className, ...props }, ref) => (
-  <SelectPrimitive.ScrollUpButton
-    ref={ref}
-    className={cn('flex cursor-default items-center justify-center py-1', className)}
-    {...props}
-  >
-    <ChevronUp className="h-4 w-4" />
-  </SelectPrimitive.ScrollUpButton>
-));
-SelectScrollUpButton.displayName = SelectPrimitive.ScrollUpButton.displayName;
-
-export { SelectScrollUpButton };
