--- packages/webapp-libs/webapp-core/src/components/forms/select/selectSeparator/index.ts
@@ -1,1 +1,0 @@
-export { SelectSeparator } from './selectSeparator.component';
