--- packages/webapp-libs/webapp-core/src/components/forms/select/selectTrigger/index.ts
@@ -1,1 +1,0 @@
-export { SelectTrigger } from './selectTrigger.component';
